Move APIs constants to new file
