+++ src/applications/appeals/995/tests/995-downtime-notification.cypress.spec.js
@@ -1,4 +1,4 @@
-import { CONTESTABLE_ISSUES_API } from '../constants';
+import { CONTESTABLE_ISSUES_API } from '../constants/apis';
 
 import mockData from './fixtures/data/minimal-test.json';
 import mockInProgress from './fixtures/mocks/in-progress-forms.json';
